requirements.txt
numpy==1.24.1
pandas==1.5.2
tsfresh==0.19.0
scikit-learn==1.2.0
seaborn==0.12.0
jupyterlab==3.4.8
etherscan==0.3.0
polygonscan-python==1.0.2
shroomdk==1.0.2
ipykernel==6.20.1